+++ src/main/java/com/library/controller/UpdateBookServlet.java
@@ -31,7 +31,7 @@ public class UpdateBookServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        long id = Long.parseLong(req.getParameter("id"));
+        int id = Integer.parseInt(req.getParameter("id"));
         BookDTO book = bookService.findById(id);
         req.setAttribute("book", book);
         req.getRequestDispatcher("/view/updateForm.jsp").forward(req, resp);
